# api/api.yaml
openapi: 3.0.0
info:
  title: ZidiBackend API
  version: 1.0.0
  description: API for managing organizations and users in ZidiBackend
paths:
...
  /users/email:
    get:
      summary: Get a user by email
      operationId: GetUserByEmail
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /auth/login:
    post:
      summary: User login
      operationId: LoginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT access token
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          description: Unauthorized - Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /campaigns:
    post:
      summary: Create a new campaign
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Campaign'
      responses:
        '201':
          description: Campaign created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          description: Invalid input

    get:
      summary: Get all campaigns
      operationId: getAllCampaigns
      responses:
        '200':
          description: A list of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
  /customers:
    post:
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid input

    get:
      summary: Get all customers
      operationId: getAllCustomers
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
  /customers/organization:
      get:
        summary: Get all customers for a specific organization
        operationId: getCustomersByOrganization
        parameters:
          - name: organizationId
            in: query
            required: true
            schema:
              type: string
              format: uuid
        responses:
          '200':
            description: List of customers for the specified organization
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Customer'
          '400':
            description: Missing or invalid organizationId
          '404':
            description: No customers found for this organization
  /customers/{id}:
      get:
        summary: Get a customer by ID
        operationId: getCustomerById
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string
              format: uuid
        responses:
          '200':
            description: Customer found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Customer'
          '404':
            description: Customer not found

      put:
        summary: Update a customer by ID
        operationId: updateCustomer
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string
              format: uuid
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        responses:
          '200':
            description: Customer updated successfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Customer'
          '404':
            description: Customer not found

      delete:
        summary: Delete a customer by ID
        operationId: deleteCustomer
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string
              format: uuid
        responses:
          '204':
            description: Customer deleted successfully
          '404':
            description: Customer not found
  /campaigns/organization:
      get:
        summary: Get all campaigns for a specific organization
        operationId: getCampaignsByOrganization
        parameters:
          - name: organizationId
            in: query
            required: true
            schema:
              type: string
              format: uuid
        responses:
          '200':
            description: List of campaigns for the specified organization
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Campaign'
          '400':
            description: Missing or invalid organizationId
          '404':
            description: No campaigns found for this organization
  /campaigns/{id}:
      get:
        summary: Get a campaign by ID
        operationId: getCampaignById
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string
              format: uuid
        responses:
          '200':
            description: Campaign found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Campaign'
          '404':
            description: Campaign not found

      put:
        summary: Update a campaign by ID
        operationId: updateCampaign
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string
              format: uuid
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        responses:
          '200':
            description: Campaign updated successfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Campaign'
          '404':
            description: Campaign not found

      delete:
        summary: Delete a campaign by ID
        operationId: deleteCampaign
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string
              format: uuid
        responses:
          '204':
            description: Campaign deleted successfully
          '404':
            description: Campaign not found
...
components:
  schemas:
...
    Campaign:
      type: object
      required:
        - id
        - campaignName
        - couponId
        - characterType
        - couponLength
        - welcomeMessage
        - questionNumber
        - amount
        - status
        - organizationId
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        campaignName:
          type: string
          example: "Summer Promo"
        couponId:
          type: string
          example: "Alphanumeric"
        characterType:
          type: string
          example: "Alphanumeric"
        couponLength:
          type: integer
          example: 10
        welcomeMessage:
          type: string
          example: "Welcome to our summer campaign!"
        questionNumber:
          type: integer
          example: 5
        amount:
          type: number
          format: float
          example: 100.50
        status:
          type: string
          example: "active"
        organizationId:
          type: string
          format: uuid
          example: "d63e4c28-f8d2-4c39-b3bb-92cd5b998a21"
    User:
      type: object
      required:
        - id
        - firstname
        - lastname
        - email
        - address
        - password
        - role
        - organization_id
      properties:
        id:
          type: string
          format: uuid
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        password:
          type: string
          format: password
        role:
          type: string
        organization_id:
          type: string
          format: uuid
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    Customer:
      type: object
      required:
        - id
        - firstName
        - lastName
        - phone
        - email
        - feedback
        - network
        - amount
        - status
        - organizationId
        - campaignId
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        phone:
          type: string
          example: "+1234567890"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        feedback:
          type: string
          example: "Great service!"
        network:
          type: string
          example: "MTN"
        amount:
          type: number
          format: float
          example: 100.50
        status:
          type: string
          example: "active"
        organizationId:
          type: string
          format: uuid
          example: "d63e4c28-f8d2-4c39-b3bb-92cd5b998a21"
        campaignId:
          type: string
          format: uuid
          example: "mka98yd6-hh72-kj86-b3bb-92cd5b998a21"
...
    Error:
      type: object
      required:
        - message
        - error_code
      properties:
        error_code:
          type: string
        message:
          type: string
  responses:
    BadRequestError:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"